ingestion/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
dotenv==0.9.9
executing==2.2.1
h11==0.16.0
h2==4.3.0
h3==4.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
postgrest==1.1.1
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
realtime==2.7.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
